Apply dust after decoration
This will cause it to run slower in biomes with dust, due to needing an extra vm:get_data() and vm:set_data(), but prevents the dust nodes from interfering with decorations (e.g. snow preventing tree growth). In native mapgens the dust is applied last, falling on top of the highest nodes.